pyproject.toml
[tool.poetry]
name = "testfligoo-data-pipeline"
version = "0.1.0"
description = "Dockerized data pipeline with Airflow, Postgres, Pandas, and Great Expectations"
authors = ["Julian Ignacio Fried julianignaciofried@gmail.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^1.5.3"
great-expectations = "^0.17.16"
psycopg2-binary = "^2.9.9"
apache-airflow = "2.7.3"
apache-airflow-providers-postgres = "^5.7.0"
sqlalchemy = ">=1.4.28,<2.0"
flask-session = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pylint = "^3.0.2"
mypy = "^1.7.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88